rename "archivist" package to "historyarchive"
