+++ tools/stellar-archivist/historyarchive/range_test.go
@@ -2,7 +2,7 @@
 // under the Apache License, Version 2.0. See the COPYING file at the root
 // of this distribution or at http://www.apache.org/licenses/LICENSE-2.0
 
-package archivist
+package historyarchive
 
 import (
 	"testing"
